test: added block scanning qa tool for scraping test data (#429)
A simple bun/typescript tool that uses indexer to scan a chain for blocks, transactions, etc.
Useful to populate test data and target the indexer subscription part.

The scanner checks if the indexer is ready on the selected target environment and starts a
stream of blocks, extracting transactions (system and regular) and storing them into a file.

This is useful for debugging as well as feeding test data to the QA indexer tests.

---------

Signed-off-by: Giuseppe Salvatore <giuseppe.salvatore@shielded.io>